Wrapping up

+++ backend/src/api.py
@@ -36,11 +36,11 @@ def drinks_short():
         success=True,
         drinks=[drink.short() for drink in drinks],
     )),
-            200) if drinks else (jsonify({}), 404)
+            200)
 
 
-# @requires_auth('get:drinks-detail')
 @app.route('/drinks-detail')
+@requires_auth('get:drinks-detail')
 def drinks_long():
     """
     GET /drinks-detail
@@ -54,31 +54,12 @@ def drinks_long():
         success=True,
         drinks=[drink.long() for drink in drinks],
     )),
-            200) if drinks else (jsonify({}), 404)
+            200)
 
 
-# @requires_auth('post:drinks')
-@app.route('/drinks', methods=["POST"])
-def create_drink():
-    """
-    POST /drinks
-        it should create a new row in the drinks table
-        it should require the 'post:drinks' permission
-        it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the newly created drink
-        or appropriate status code indicating reason for failure
-    """
-    req = request.json.get
-    drink = Drink(title=req('title'), recipe=req('recipe'))
-    return (jsonify(dict(
-        success=True,
-        drinks=[drink],
-    )), 200) if drink else (jsonify({}), 406)
-
-
-# @requires_auth('patch:drinks')
 @app.route('/drinks/<int:id>', methods=["PATCH"])
-def patch_drink():
+@requires_auth('patch:drinks')
+def patch_drink(id):
     """
     PATCH /drinks/<id>
         where <id> is the existing model id
@@ -90,19 +71,44 @@ def patch_drink():
         or appropriate status code indicating reason for failure
     """
     drink = Drink.query.filter(Drink.id == id).one_or_none()
+    if not drink:
+        abort(404)
     for att in request.json:
         setattr(drink, att, request.json.get(att))
     drink.update()
 
     return (jsonify(dict(
         success=True,
-        drinks=[drink],
+        drinks=[drink.long()],
     )), 200) if drink else (jsonify({}), 405)
 
 
-# @requires_auth('delete:drinks')
+@app.route('/drinks', methods=["POST"])
+@requires_auth('post:drinks')
+def create_drink():
+    """
+    POST /drinks
+        it should create a new row in the drinks table
+        it should require the 'post:drinks' permission
+        it should contain the drink.long() data representation
+    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the newly created drink
+        or appropriate status code indicating reason for failure
+    """
+    try:
+        req = request.json.get
+    except AttributeError:
+        abort(500)
+    drink = Drink(title=req('title'), recipe=json.dumps(request.json['recipe']))
+    drink.insert()
+    return (jsonify(dict(
+        success=True,
+        drinks=[drink.long()],
+    )), 200)
+
+
 @app.route('/drinks/<int:id>', methods=["DELETE"])
-def delete_drink():
+@requires_auth('delete:drinks')
+def delete_drink(id):
     """
     DELETE /drinks/<id>
         where <id> is the existing model id
@@ -113,13 +119,15 @@ def delete_drink():
         or appropriate status code indicating reason for failure
     """
     drink = Drink.query.filter(Drink.id == id).one_or_none()
+    if not drink:
+        abort(404)
     id = drink.id
     drink.delete()
 
     return (jsonify(dict(
         success=True,
         delete=id,
-    )), 200) if drink else (jsonify({}), 405)
+    )), 200)
 
 
 # # Error Handling
@@ -156,6 +164,15 @@ def not_found(error):
         "error": 404,
         "message": "not found"
     }), 404
+
+
+@app.errorhandler(500)
+def not_found(error):
+    return jsonify({
+        "success": False,
+        "error": 500,
+        "message": "the server could not complete the request"
+    }), 500
 
 
 @app.errorhandler(AuthError)
